@@ -1,4 +1,3 @@
-
 buildscript {
     repositories {
         google()
@@ -6,11 +5,12 @@ buildscript {
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.2'
+        classpath 'com.android.tools.build:gradle:3.6.2'
     }
 }
 
 apply plugin: 'com.android.application'
+apply plugin: 'com.google.gms.google-services'
 
 repositories {
     google()
@@ -20,15 +20,13 @@ repositories {
 dependencies {
 
 
-        implementation "com.android.support:support-v4:28.0.0"
-        implementation "com.android.support:support-v13:28.0.0"
-        implementation "com.android.support:cardview-v7:28.0.0"
-        implementation "com.android.support:appcompat-v7:28.0.0"
-        implementation "com.android.support:support-compat:28.0.0"
-
-
-
-
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.core:core:1.0.0'
+    implementation 'com.google.firebase:firebase-messaging:20.1.5'
+    implementation 'com.google.firebase:firebase-analytics:17.2.2'
 
 }
 
@@ -36,16 +34,16 @@ dependencies {
 // keep boilerplate and common code separate from
 // the main sample code.
 List<String> dirs = [
-    'main',     // main sample code; look here for the interesting stuff.
-    'common',   // components that are reused by multiple samples
-    'template'] // boilerplate code that is generated by the sample template process
+        'main',     // main sample code; look here for the interesting stuff.
+        'common',   // components that are reused by multiple samples
+        'template'] // boilerplate code that is generated by the sample template process
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 28
+        targetSdkVersion 29
     }
 
     compileOptions {
@@ -64,5 +62,6 @@ android {
         androidTest.java.srcDirs = ['tests/src']
 
     }
+    buildToolsVersion = '29.0.3'
 
 }
